Refactor page generation (#335)
* Remove unused catalogs from page context

* Remove redundant folder check

* Replace single MDX page generation with per-type

* Update explainers template

* Refactor page generation

This moves us much, much closer to having shared logic for international routes by refactoring existing methods to use a more similar scheme. It first creates an object keyed by route, with each translation in an array. This gives us a nice point to abstract the various pages, though pagination is still an unknown.

This also drops the `/en/*` pages that were being created. They're just cruft atm, and might not be beneficial in the future. Replace it with a redirect to top level pages.

* Rename build helpers that don't create pages themselves

* Refactor to share i18n route logic

* Unify page creation fragments

* Group by route in i18n page helper

* Update comment for non-standard i18n