Sync up api ref routing and sidebar
These were using similar (but not identical) methods to track which section was active, with the router tracking the sections themselves, and the sidebar tracking headings. Because of this the URL and "active section" in the sidebar frequently got out of sync. This change overrides the side nav position tracking so that it's all handled by ScrollRouter on API Reference.