Polish scroll position tracking (#116)
* Remove intersection observer polyfill

It's pretty well supported, basically only IE doens't have support.

* Track bounding div, not first child

* Fix active section detection logic

* Only query for content node once

* Change TrackedElement to expect an identifier

By requesting it as a prop instead of peeling it off the first child, we get more flexibility. With the changes to tracking logic, this simplifies how the logic is working

* Move sidenav provider out of layout base

* Sync up api ref routing and sidebar

These were using similar (but not identical) methods to track which section was active, with the router tracking the sections themselves, and the sidebar tracking headings. Because of this the URL and "active section" in the sidebar frequently got out of sync. This change overrides the side nav position tracking so that it's all handled by ScrollRouter on API Reference.

* Suppress attribute warning