Remove /api from sitemap
This will hopefully fix some SEO issues. Some search terms pull up /api as the destination, which is specifically what we want to avoid. The intended URL appears to always be there as well, so by removing this from the sitemap I think Google will correctly point to the right destination