Fix GA sessions, remove /api from sitemap, update landing page (#176)
* Remove /api from sitemap

This will hopefully fix some SEO issues. Some search terms pull up /api as the destination, which is specifically what we want to avoid. The intended URL appears to always be there as well, so by removing this from the sitemap I think Google will correctly point to the right destination

* Use autolinker for cross-domain google analytics

* Show developers intro page from stellar.org as top-level page

* Whoops