Update decentralized-exchange.mdx (#225)
* Update decentralized-exchange.mdx

Trading terminology will be confusing because "offers" is typically associated with "sell orders" and synonymous with "asks." I know the terminology in the code is "offers" and changing "offers" to "orders" here might be confusing.

(minor comment): Using "consumed" instead of "taken" because when you "take an offer" you are typically buying an ask (alternatively, selling a bid is "hitting a bid")...just conventional trading lingo

"Prices are specified as a `{numerator, denominator}` pair, with both components of the fraction represented as 32 bit signed integers. When expressing a price of "Asset A in terms of Asset B", the amount of B is always the numerator, and A is always the denominator. "
 - this is backwards...with currency pairs, "Asset A in terms of Asset B" usually means Asset B is the quote asset, and A is the base asset. These are always quoted as base/quote (so A is numerator and B is denominator). When you're talking about buying/selling an fx pair, you're generally talking in base terms (i.e. I'm buying/selling yen... which would be represented yen/(quoteasset).

* Update decentralized-exchange.mdx