Issue #360: Updating examples to use simplified accounts
I've used `sed` to go through all the mdx files in the `content/`
directory and changed the Go code samples to use the new, simplified syntax.

I'd love some feedback, as I've never used Go. I'm not sure if there are
other places to look for these code samples in the repo, as well.

Signed-off-by: Elliot J. Voris <elliot@voris.me>