Update command to install postcss 2 (#412)
At the time of this commit, Stencil 2 is the default version when
`npm init stencil` is executed. By following the tutorial, the
installation of the SCSS tools will fail when installing postcss v1.

This change adapts the installation command to lock on version 2 as
major for the postcss library, avoiding dependency tree problems.